+++ changeLog.md
@@ -3,6 +3,7 @@
 ## v0.0.11:
  - revised v0.0.10 declarativeContent is required to start the popup pageAction
  - fixed bug with sorting
+ - fixed bug in changing name function
 
 ## v0.0.10:
  - removed declarativeContent permission